modelクラスをファイル分割
