+++ functions/model/store_detail_list.py
@@ -1,0 +1,3 @@
+class StoreDetailList:
+    def __init__(self, store_detail_obj_list) -> None:
+        self.store_detail_obj_list = store_detail_obj_list
